Get rid of token expiration
It might be more secure to have expiring tokens, but it currently
seems unreasonable to burden users with the necessity to re-login,
even once a month.
Instead, if something goes really wrong, we can always reintroduce
expiration, while expiring all existing tokens simultaneously is
as easy as changing JWT_SECRET on the server.